+++ src/dashboardWebView/components/Media/Media.tsx
@@ -272,6 +272,8 @@ export const Media: React.FunctionComponent<IMediaProps> = (
         version={settings?.versionInfo}
         isBacker={settings?.isBacker}
       />
+
+      <img className='hidden' src="https://api.visitorbadge.io/api/visitors?path=https%3A%2F%2Ffrontmatter.codes%2Fmetrics%2Fdashboards&slug=media" alt="Media metrics" />
     </PageLayout>
   );
 };