+++ src/dashboardWebView/components/SnippetsView/Snippets.tsx
@@ -180,6 +180,8 @@ export const Snippets: React.FunctionComponent<ISnippetsProps> = (
         version={settings?.versionInfo}
         isBacker={settings?.isBacker}
       />
+
+      <img className='hidden' src="https://api.visitorbadge.io/api/visitors?path=https%3A%2F%2Ffrontmatter.codes%2Fmetrics%2Fdashboards&slug=snippets" alt="Snippets metrics" />
     </PageLayout>
   );
 };