+++ CHANGELOG.md
@@ -6,11 +6,13 @@
 
 - External UI script support for dashboards
 - Visual Studio Code Theming support for the dashboards
+- Front matter AI 🤖
 
 > **Info**: To enable the experimental features you need to set the `frontMatter.experimental` setting to `true`.
 
 ### ✨ New features
 
+- [#530](https://github.com/estruyf/vscode-front-matter/issues/530): Implementation of the Front Matter AI 🤖 powered by [mendable.ai](https://mendable.ai)
 - [#513](https://github.com/estruyf/vscode-front-matter/issues/513): Added support for external UI scripts to add custom HTML on the dashboard elements
 
 ### 🎨 Enhancements